# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "HelpDesk AI API",
        "version": "1.0.0",
        "description": "Auto-generated from routes/api.php. Auth: Sanctum Bearer token or scoped API key (X-API-Key). Envelope: {success, data, message}."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum|API-Key"
            }
        }
    },
    "paths": {
...
        "/api/analytics/summary": {
            "get": {
                "operationId": "api.analytics.summary",
                "summary": "AnalyticsController@summary",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Analytics"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK \u2014 envelope {success, data, message}"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    }
}